Reorganizing output formats into xxporters package
